fix: moves group logic to group manager :/
