+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Operations/Receive/RhinoHostObjectBuilder.cs
@@ -1,6 +1,5 @@
 using Rhino;
 using Rhino.DocObjects;
-using Rhino.DocObjects.Custom;
 using Rhino.Geometry;
 using Speckle.Connectors.Rhino.HostApp;
 using Speckle.Connectors.Utils.Builders;
@@ -30,6 +29,7 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
   private readonly RhinoLayerManager _layerManager;
   private readonly RhinoMaterialManager _materialManager;
   private readonly RhinoColorManager _colorManager;
+  private readonly RhinoGroupManager _groupManager;
   private readonly ISyncToThread _syncToThread;
 
   public RhinoHostObjectBuilder(
@@ -40,6 +40,7 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     RhinoInstanceObjectsManager instanceObjectsManager,
     RhinoMaterialManager materialManager,
     RhinoColorManager colorManager,
+    RhinoGroupManager groupManager,
     ISyncToThread syncToThread
   )
   {
@@ -50,6 +51,7 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     _instanceObjectsManager = instanceObjectsManager;
     _materialManager = materialManager;
     _colorManager = colorManager;
+    _groupManager = groupManager;
     _syncToThread = syncToThread;
   }
 
@@ -218,29 +220,11 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     // Stage 3: Groups
     if (groupProxies is not null)
     {
-      using (var _ = SpeckleActivityFactory.Start("Converting groups"))
-      {
-        BakeGroups(groupProxies, applicationIdMap, baseLayerName);
-      }
+      _groupManager.BakeGroups(groupProxies, applicationIdMap, baseLayerName);
     }
 
     // Stage 4: Return
     return new(bakedObjectIds, conversionResults);
-  }
-
-  private void BakeGroups(
-    List<GroupProxy> groupProxies,
-    Dictionary<string, List<string>> applicationIdMap,
-    string baseLayerName
-  )
-  {
-    using var _ = SpeckleActivityFactory.Start();
-    foreach (GroupProxy groupProxy in groupProxies.OrderBy(g => g.objects.Count))
-    {
-      var appIds = groupProxy.objects.SelectMany(oldObjId => applicationIdMap[oldObjId]).Select(id => new Guid(id));
-      var groupName = (groupProxy.name ?? "No Name Group") + $" ({baseLayerName})";
-      _contextStack.Current.Document.Groups.Add(groupName, appIds);
-    }
   }
 
   private void PreReceiveDeepClean(string baseLayerName)
@@ -271,15 +255,8 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
       }
     }
 
-    // Cleans up any previously received groups
-    for (int i = _contextStack.Current.Document.Groups.Count; i >= 0; i--)
-    {
-      var group = _contextStack.Current.Document.Groups.FindIndex(i);
-      if (group is { Name: not null } && group.Name.Contains(baseLayerName))
-      {
-        _contextStack.Current.Document.Groups.Delete(i);
-      }
-    }
+    // Cleans up any previously received group
+    _groupManager.PurgeGroups(baseLayerName);
   }
 
   private IReadOnlyList<string> HandleConversionResult(
